Namespace change for MediatR
