+++ Src/RCommon.Mediator.Mediatr/Behaviors/ValidatorBehavior.cs
@@ -9,7 +9,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace RCommon.Mediatr.Behaviors
+namespace RCommon.Mediator.MediatR.Behaviors
 {
     public class ValidatorBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
         where TRequest : IRequest<TResponse>
