+++ Src/RCommon.Mediator.Mediatr/IMediatRConfiguration.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RCommon.Mediatr
+namespace RCommon.Mediator.MediatR
 {
     public interface IMediatrConfiguration : IMediatorConfiguration
     {
