+++ Src/RCommon.Mediator.Mediatr/MediatRConfiguration.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RCommon.Mediatr
+namespace RCommon.Mediator.MediatR
 {
     public class MediatrConfiguration : IMediatrConfiguration
     {
