+++ Src/RCommon.Mediator.Mediatr/RCommon.Mediator.MediatR.csproj
